+++ frontend/src/components/teamer/pages/TeamerProfilePage.tsx
@@ -36,7 +36,7 @@ import ChangeRoleTitleModal from '../../admin/modals/ChangeRoleTitleModal';
 import DeleteAccountModal from '../../shared/DeleteAccountModal';
 import SpiritFooter from '../../shared/SpiritFooter';
 import TeamerOnboardingModal from '../modals/TeamerOnboardingModal';
-import TeamerUpdateWalkthroughModal from '../modals/TeamerUpdateWalkthroughModal';
+import TeamerUpdate211WalkthroughModal from '../modals/TeamerUpdate211WalkthroughModal';
 import WrappedModal from '../../wrapped/WrappedModal';
 import type { WrappedHistoryEntry } from '../../../types/wrapped';
 import LoadingSpinner from '../../common/LoadingSpinner';
@@ -648,7 +648,7 @@ const TeamerProfilePage: React.FC = () => {
       )}
 
       {showUpdateWalkthrough && (
-        <TeamerUpdateWalkthroughModal onClose={() => setShowUpdateWalkthrough(false)} />
+        <TeamerUpdate211WalkthroughModal onClose={() => setShowUpdateWalkthrough(false)} />
       )}
 
       {showMitmachenErklaerung && (